requirements.txt
altgraph==0.17.4
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mutagen==1.47.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
proglog==0.1.10
pycryptodomex==3.20.0
pydub==0.25.1
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
PySide6==6.7.0
PySide6_Addons==6.7.0
PySide6_Essentials==6.7.0
requests==2.32.2
shiboken6==6.7.0
tqdm==4.66.2
urllib3==2.2.1
websockets==12.0
yt-dlp==2024.5.20.232721.dev0
